naive parsing and math in Haskell; no math any more in gnuplot; 2 new scripts
Warning: the old plot scripts don't work with this new Haskell code. TODO.

Gnuplot can't do differential quotients, so the math is done in Haskell.
But the simple differential quotients in this version
are not a good idea: the plots are often unreadable
(hundreds of narrow peaks) and very unstable over the density
of sample points. Perhaps the quotients should be taken over fixed small
time intervals, not the intervals between consequtive sample points..