add combined poll size graphs
Indeed they help verify the correctness of the aggregation. It seems OK,
but the max/min look a little funny around the original graph,
sometimes seems a bit too distant, so perhaps an extra rule with look ahead
is needed that relaxes the interval inspection. It shouldn't result
in a solid blob, but it will compicate the code.